+++ packages/zenoh/lib/src/zenoh.dart
@@ -8,7 +8,6 @@ import 'package:ffi/ffi.dart';
 import 'config.dart';
 import 'hello.dart';
 import 'id.dart';
-import 'bindings.dart' as ffi_bindings;
 import 'native_lib.dart';
 import 'whatami.dart';
 
@@ -35,10 +34,9 @@ class Zenoh {
   ///
   /// See: https://docs.rs/env_logger/latest/env_logger/
   static void initLog(String fallback) {
-    ensureInitialized();
     final cStr = fallback.toNativeUtf8();
     try {
-      ffi_bindings.zd_init_log(cStr.cast<Char>());
+      bindings.zd_init_log(cStr.cast<Char>());
     } finally {
       calloc.free(cStr);
     }
@@ -107,7 +105,7 @@ class Zenoh {
     final cfgPtr = configAddr != 0
         ? Pointer<Void>.fromAddress(configAddr)
         : nullptr;
-    ffi_bindings.zd_scout(cfgPtr.cast(), nativePort, timeoutMs, what);
+    bindings.zd_scout(cfgPtr.cast(), nativePort, timeoutMs, what);
 
     return completer.future;
   }